fix(schema): Change message_index from Int8 to Int4 (i64 to i32)

+++ src/core/shared/models/core.rs
@@ -136,7 +136,7 @@ pub struct MessageHistory {
     pub role: i32,
     pub content_encrypted: String,
     pub message_type: i32,
-    pub message_index: i64,
+    pub message_index: i32,
     pub created_at: DateTime<Utc>,
 }
